Array.prototype.indexOf/lastIndexOf/includes: handle Infinity fromIndex
parseFromIndex (native path) and the typed _from variants used int64_t
for fromIndex, so +/-Infinity became INT64_MIN (UB) after fptosi and
produced wrong results:
  [true].indexOf(true, -Infinity) returned -1 (spec: 0)
  [true].lastIndexOf(true, -Infinity) returned 0 (spec: -1)

Per ES spec (ToIntegerOrInfinity):
  indexOf:     -Infinity -> 0;          +Infinity -> miss (-1)
  lastIndexOf: -Infinity -> miss (-1);  +Infinity -> len-1
  includes:    -Infinity -> 0;          +Infinity -> miss (false)
  NaN -> 0 in all cases (already handled).

Runtime signatures changed from int64_t fromIndex to double fromIndex
so Infinity survives the call. LoweringRegistry now passes f64 instead
of i64. Native parseFromIndex gets matching branches.

test262 3k sample: 862 -> 881 (+19 tests).
"no message" cluster: 84 -> 72 (-12 here, -7 elsewhere from earlier fixes).
No regressions in golden-IR (265/277) or node (293/297).

+++ src/runtime/include/TsArray.h
@@ -134,11 +134,11 @@ extern "C" {
     void* ts_array_flat(void* arr, int64_t depth);
     void* ts_array_flatMap(void* arr, void* callback, void* thisArg);
     int64_t ts_array_indexOf(void* arr, int64_t value);
-    int64_t ts_array_indexOf_from(void* arr, int64_t value, int64_t fromIndex);
+    int64_t ts_array_indexOf_from(void* arr, int64_t value, double fromIndex);
     int64_t ts_array_lastIndexOf(void* arr, int64_t value);
-    int64_t ts_array_lastIndexOf_from(void* arr, int64_t value, int64_t fromIndex);
+    int64_t ts_array_lastIndexOf_from(void* arr, int64_t value, double fromIndex);
     bool ts_array_includes(void* arr, int64_t value);
-    bool ts_array_includes_from(void* arr, int64_t value, int64_t fromIndex);
+    bool ts_array_includes_from(void* arr, int64_t value, double fromIndex);
     void* ts_array_at(void* arr, int64_t index);
     void* ts_array_join(void* arr, void* separator);
     void* ts_array_concat(void* arr, void* other);
